Implement chunked response (#15)
Instead of sending the file all at once, now the httpgw writes it chunk by chunk using the chunked encoding option.
To avoid bottlenecking the gateway, when a chunk is sent, it is eliminated from the gateway's aux structures, meaning that the gateway is not limited by the ram size of the computer in which it is running.

+++ HttpGw/src/main/java/FSChunk/ChunkThread.java
@@ -14,7 +14,7 @@ import java.util.concurrent.locks.ReentrantLock;
 public class ChunkThread extends Thread {
     private ArrayList<Integer> offsets;
     private String fileName;
-    private int fileSize;
+    private long fileSize;
     private int packetSize;
     private boolean last;
     private Map<Integer, MyPair<Condition,byte[]>> responseStruct;
@@ -24,7 +24,7 @@ public class ChunkThread extends Thread {
 
     public ChunkThread(ArrayList<Integer> offsets,
                        String fileName,
-                       int fileSize,
+                       long fileSize,
                        int packetSize,
                        boolean last,
                        Map<Integer, MyPair<Condition,byte[]>> responseStruct,
@@ -53,7 +53,7 @@ public class ChunkThread extends Thread {
         // Capture same length packets
         for(int i = 0; i < equalPackets;i++) {
             try {
-                fileChunk = worker.getFile(offsets.get(i) * packetSize,packetSize);
+                fileChunk = worker.getFile((long) offsets.get(i) * packetSize,packetSize);
                 arrayLock.lock();
                 responseStruct.get(offsets.get(i)).setSecond(fileChunk);
                 // Wakes up the retriever thread
@@ -67,9 +67,9 @@ public class ChunkThread extends Thread {
         // Capture last different length packet, if this thread has one
         if(last) {
             int lastOffset = offsets.get(offsets.size()-1);
-            int lastPacketSize = fileSize - lastOffset * packetSize;
+            int lastPacketSize = (int) (fileSize - lastOffset * packetSize);
             try {
-                fileChunk = worker.getFile(lastOffset * packetSize,lastPacketSize);
+                fileChunk = worker.getFile((long) lastOffset * packetSize,lastPacketSize);
                 arrayLock.lock();
                 responseStruct.get(lastOffset).setSecond(fileChunk);
                 // Wakes up the retriever thread
