+++ HttpGw/src/main/java/FSChunk/FileMetaData.java
@@ -2,7 +2,7 @@ package FSChunk;
 
 public class FileMetaData {
     private final boolean exists;
-    private final int size;
+    private final long size;
     private final String type;
 
     public FileMetaData(String content) {
@@ -14,7 +14,7 @@ public class FileMetaData {
         else {
             String[] tokens = content.split(",");
             exists = Boolean.parseBoolean(tokens[0].split(":")[1]);
-            size = Integer.parseInt(tokens[1].split(":")[1]);
+            size = Long.parseLong(tokens[1].split(":")[1]);
             type = tokens[2].split(":")[1].trim();
         }
     }
@@ -23,7 +23,7 @@ public class FileMetaData {
         return exists;
     }
 
-    public int getSize() {
+    public long getSize() {
         return size;
     }
